Use HTTPS URLs for submodules instead of SSH URLs
When pushing to a Git server which cannot connect to Github via SSH (or needs to be granted an SSH key to do so), the submodules can't be downloaded.

This PR is replacing the SSH URLs which need an SSH key, with ``https://`` URLs. We could also use plain ``git://`` URLs in that case but HTTPS seems better.